+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * ReflectionPropertyExtractor.php
 * Fix: Archive:Load X more posts on is_home() first page
+* Gutenberg: disable empty paragraph blocks
 
 ## 0.2.0 (2022-02-15)